Experimenting

+++ Game/DungeonLife/Assets/Scripts/WeaponConfig.cs
@@ -6,9 +6,9 @@ public class WeaponConfig : ScriptableObject
     public float damage;
     public float speed;
 
-    public void AddAttack(float value) 
+    public void AddAttack(int value) 
         => speed += value;
 
-    public void AddDamage(float value) 
+    public void AddDamage(int value) 
         => damage += value;
 }
